+++ packages/plugin-inline-script-content/inline-script-content.js
@@ -126,10 +126,9 @@ module.exports = (doc = document, win = window) => ({
     })
 
     function __traceOriginalScript (fn, callbackAccessor, alsoCallOriginal = false) {
-      return function () {
+      return function (...args) {
         // this is required for removeEventListener to remove anything added with
         // addEventListener before the functions started being wrapped by Bugsnag
-        const args = [].slice.call(arguments)
         try {
           const cba = callbackAccessor(args)
           const cb = cba.get()
@@ -142,14 +141,14 @@ module.exports = (doc = document, win = window) => ({
             // this function mustn't be annonymous due to a bug in the stack
             // generation logic, meaning it gets tripped up
             // see: https://github.com/stacktracejs/stack-generator/issues/6
-            cb.__trace__ = function __trace__ () {
+            cb.__trace__ = function __trace__ (...cbArgs) {
               // set the script that called this function
               updateLastScript(script)
               // immediately unset the currentScript synchronously below, however
               // if this cb throws an error the line after will not get run so schedule
               // an almost-immediate aysnc update too
               _setTimeout(function () { updateLastScript(null) }, 0)
-              const ret = cb.apply(this, arguments)
+              const ret = cb.apply(this, cbArgs)
               updateLastScript(null)
               return ret
             }